Add .editorconfig file
"EditorConfig helps maintain consistent coding styles for multiple
developers working on the same project across various editors and IDEs."
https://editorconfig.org/

Basically the idea here is that editors with an editorconfig plugin use
this file to consistently and automatically apply settings across all
platforms and editors. Super useful for teams with a diverse set of dev
machines.
The plugin for vscode is here:
https://marketplace.visualstudio.com/items?itemName=EditorConfig.EditorConfig