Manage python dependencies with pdm
This commit sets up the python server to manage dependencies with
[pdm](https://pdm.fming.dev/latest/). pdm supports the new [PEP
582](https://peps.python.org/pep-0582/) which allows dependencies to be
installed directly in the project directory obviating the need for
virtualenvs and greatly simplying the development and onboarding
process.

pdm does allow exporting a requirements.txt file, which is what the
existing docker-compose setup expects, but follow up work would/will
make adjustments to take advantage of what pdm brings to simplify that
setup.