Use a proxy for local dev
This commit adds the envoy proxy to the development setup so that the
connectivity patterns between the services resemble the eventual production
environment. In production both services will be served from a single "origin":
the combination of domain + port.
This new setup gets that working locally:
localhost:3000 > frontend
localhost:3000/graphql > api